refactor: reorganize project structure for better separation of concerns
